// src/app/model/products.ts
export class Junta{
  /*
    pName: string;
    desc: string;
    price: string
    id?: string;
  */
    nominal: string;
    nominal1: string;
    lineaOSistema: string;
    especificacion: string;
    schedule: string;
    diam_inch_contabilizadas: string;
    factor_pulgadas_diametrales: string;
    pulgadas_diametrales: string;
    tipo_extremos: string;
    tipo_material: string;
    material: string;
    id?: string;

}